Make the code less fragile.
I changed aref_voltage from 3.3 to 5 and suddenly my temperature was zero. The problem is that when aref_voltage is #defined as 3.3 then in the expression at line 178 aref_voltage is taken as a floating point value and dividing it by 1024 gives the correct value. When I #defined it as 5 it happened to be treated as an integer and then 5/1024 is zero.